+++ phytreon/comparative/__init__.py
@@ -11,7 +11,9 @@ from .community import (
     patristic_distances, mpd, mntd, ses_mpd, ses_mntd,
     beta_mntd, beta_nti, permanova, mantel,
 )
-from .signal import phylo_vcv, blomberg_k, pagels_lambda, pgls
+from .signal import phylo_vcv, blomberg_k, pagels_lambda, pgls, fritz_purvis_d
+from .models import (fit_continuous, compare_continuous_models, phylo_pca,
+                     MODELS as CONTINUOUS_MODELS)
 
 __all__ = [
     "ace_parsimony", "ace_ml", "ace_continuous", "stochastic_map",
@@ -19,5 +21,7 @@ __all__ = [
     "unweighted_unifrac", "weighted_unifrac", "unifrac_matrix",
     "patristic_distances", "mpd", "mntd", "ses_mpd", "ses_mntd",
     "beta_mntd", "beta_nti", "permanova", "mantel",
-    "phylo_vcv", "blomberg_k", "pagels_lambda", "pgls",
+    "phylo_vcv", "blomberg_k", "pagels_lambda", "pgls", "fritz_purvis_d",
+    "fit_continuous", "compare_continuous_models", "phylo_pca",
+    "CONTINUOUS_MODELS",
 ]